Various Fixes
Added to_string function for CustomPoint and Polyline
Replaced the intersection function created earlier with one created by ChatGPT
The extend function is now partially working for the first iteration, but the second one breaks the code. Might be related to the reason for the crash when trying to save after using the extend function. Maybe the x and y values are of different length as the error is: IndexError: list index out of range